refactor: position type inference and extract Vote schema to seperated file
