+++ src/index.ts
@@ -6,6 +6,7 @@ import { AuthService } from "./services/auth.service";
 import { html } from "@elysiajs/html";
 import { devRoutes } from "./routes/dev/route";
 import { electionInfo } from "./constants";
+import { Vote } from "./schemas/election.schema";
 
 const swaggerOptions = (): ElysiaSwaggerConfig<"/reference"> => {
 	return {
@@ -131,39 +132,7 @@ const app = new Elysia({
 		},
 		{
 			body: t.Object({
-				votes: t.Array(
-					t.Object({
-						candidateId: t.Union(
-							[
-								t.Number({
-									description: "Student ID of the candidate",
-									title: "Student ID",
-								}),
-								t.Literal("no-vote", {
-									title: "No Vote",
-									description:
-										"Cast no vote when multiple candidate or not express opinion on one candidate",
-								}),
-								t.Literal("disapprove", {
-									title: "Disapprove",
-									description: "Not approve when there is only one candidate",
-								}),
-							],
-							{
-								description:
-									'The student ID of the candidate or "no-vote" (cast no vote when multiple candidate) or "disapprove (when there is only one candidate)',
-								examples: [1234567823, "no-vote", "disapprove"],
-							},
-						),
-						position: t.Union([
-							...electionInfo.positions.map((position) =>
-								t.Literal(position.const, {
-									description: position.description,
-								}),
-							),
-						]),
-					}),
-				),
+				votes: t.Array(Vote),
 			}),
 			detail: {
 				description: "Submit votes from a student",
